refact(bank accounts): rename files for reducers
