+++ src/reducers/index.ts
@@ -8,8 +8,8 @@ export { default as currentOrganization }  from './current-organization';
 export { default as transaction }          from './transaction';
 export { default as transactions }         from './transactions';
 export { default as members }              from './members';
-export { default as bankAccounts }         from 'services/bank-accounts/reducers.js';
-export { default as bankAccount }          from 'services/bank-accounts/reducer.js';
+export { default as bankAccounts }         from 'services/bank-accounts/entities-reducer.js';
+export { default as bankAccount }          from 'services/bank-accounts/entity-reducer.js';
 export { default as categories }           from './categories';
 export { default as category }             from './category';
 export { default as customers }            from './customers';
